Made a drawing class, moved some files around, and cleaned up some imports
